+++ src/test/scala/atmos/retries/RetryDSLSpec.scala
@@ -38,6 +38,7 @@ class RetryDSLSpec extends FunSpec with Matchers {
   describe("RetryDSL") {
 
     it("should create retry policies by describing termination policies") {
+      neverRetry shouldEqual RetryPolicy(ImmediatelyTerminate)
       retrying shouldEqual RetryPolicy()
       retryFor { 5.attempts } shouldEqual RetryPolicy(LimitNumberOfAttempts(5))
       retryFor { 5.minutes } shouldEqual RetryPolicy(LimitAmountOfTimeSpent(5.minutes))