+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 1.1 (2013-??-??)
 ================
+ - In `atmos.retries`, added a termination policy that immediately terminates.
  - Added `atmos.utils.Logger`, a super-thin wrapper around `java.util.logging`.
  - Added `atmos.utils.Encoding`, an interface for converting arbitrary types to and from strings.
  - Added `atmos.utils.Attributes`, a template class for defining type-safe key/value pairs based on maps of strings.